+++ CHANGELOG.md
@@ -1,13 +1,19 @@
 
 # Changelog
 
+## [2.97] - 2024-11-09
+
+### Added
+- Debug support for most necessary functions with a wrapper.
+- Card for showing the node stake value currently in network.
+
 ## [2.96] - 2024-11-03
 
 ### Added
 - Add support for configuring rate limit via `rate_limit_interval` configuration value.
 
 ### Changed
-- Logging time format is changed to `isoformat()`
+- Logging time format is changed to `isoformat()`.
 - Moved configs to a separate class.
 - Clear button for deleted cards is now opaque so it doesn't block the view unless hovered.
 - Removed GMail support only and added support for any SSL/TLS capable SMTP server. **IMPORTANT: CHANGE YOUR CONFIGURATION IN `cellframe-node.cfg` FOR GMAIL**
@@ -18,7 +24,6 @@ smtp_password=<your_app_password>
 smtp_user=<your_gmail_user>
 email_use_tls=true
 ```
-
 
 ## [2.95] - 2024-10-28